requirements.txt
ccxt>=4.0.0
requests>=2.31.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio
json5>=0.9.0 